requirements.txt
Flask==2.3.2
requests
torch
deap
numpy
websocket-client
early-stopping-pytorch